requirements.txt
python-telegram-bot==20.0
python-dotenv==1.0.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0